fixing compile warnings and errors

+++ src/test/addrman_tests.cpp
@@ -91,14 +91,14 @@ BOOST_FIXTURE_TEST_SUITE(addrman_tests, BasicTestingSetup)
         CNetAddr source = ResolveIP("252.2.2.2");
 
         // Test: Does Addrman respond correctly when empty.
-        BOOST_CHECK_EQUAL(addrman.size(), 0);
+        BOOST_CHECK_EQUAL(addrman.size(), (uint64_t)0);
         CAddrInfo addr_null = addrman.Select();
         BOOST_CHECK_EQUAL(addr_null.ToString(), "[::]:0");
 
         // Test: Does Addrman::Add work as expected.
         CService addr1 = ResolveService("250.1.1.1", 8767);
         BOOST_CHECK(addrman.Add(CAddress(addr1, NODE_NONE), source));
-        BOOST_CHECK_EQUAL(addrman.size(), 1);
+        BOOST_CHECK_EQUAL(addrman.size(), (uint64_t)1);
         CAddrInfo addr_ret1 = addrman.Select();
         BOOST_CHECK_EQUAL(addr_ret1.ToString(), "250.1.1.1:8767");
 
@@ -106,7 +106,7 @@ BOOST_FIXTURE_TEST_SUITE(addrman_tests, BasicTestingSetup)
         //  Expected dup IP should not be added.
         CService addr1_dup = ResolveService("250.1.1.1", 8767);
         BOOST_CHECK(!addrman.Add(CAddress(addr1_dup, NODE_NONE), source));
-        BOOST_CHECK_EQUAL(addrman.size(), 1);
+        BOOST_CHECK_EQUAL(addrman.size(), (uint64_t)1);
 
 
         // Test: New table has one addr and we add a diff addr we should
@@ -121,7 +121,7 @@ BOOST_FIXTURE_TEST_SUITE(addrman_tests, BasicTestingSetup)
 
         // Test: AddrMan::Clear() should empty the new table.
         addrman.Clear();
-        BOOST_CHECK_EQUAL(addrman.size(), 0);
+        BOOST_CHECK_EQUAL(addrman.size(), (uint64_t)0);
         CAddrInfo addr_null2 = addrman.Select();
         BOOST_CHECK_EQUAL(addr_null2.ToString(), "[::]:0");
 
@@ -141,23 +141,23 @@ BOOST_FIXTURE_TEST_SUITE(addrman_tests, BasicTestingSetup)
 
         CNetAddr source = ResolveIP("252.2.2.2");
 
-        BOOST_CHECK_EQUAL(addrman.size(), 0);
+        BOOST_CHECK_EQUAL(addrman.size(), (uint64_t)0);
 
         // Test 7; Addr with same IP but diff port does not replace existing addr.
         CService addr1 = ResolveService("250.1.1.1", 8767);
         addrman.Add(CAddress(addr1, NODE_NONE), source);
-        BOOST_CHECK_EQUAL(addrman.size(), 1);
+        BOOST_CHECK_EQUAL(addrman.size(), (uint64_t)1);
 
         CService addr1_port = ResolveService("250.1.1.1", 8334);
         addrman.Add(CAddress(addr1_port, NODE_NONE), source);
-        BOOST_CHECK_EQUAL(addrman.size(), 1);
+        BOOST_CHECK_EQUAL(addrman.size(), (uint64_t)1);
         CAddrInfo addr_ret2 = addrman.Select();
         BOOST_CHECK_EQUAL(addr_ret2.ToString(), "250.1.1.1:8767");
 
         // Test: Add same IP but diff port to tried table, it doesn't get added.
         //  Perhaps this is not ideal behavior but it is the current behavior.
         addrman.Good(CAddress(addr1_port, NODE_NONE));
-        BOOST_CHECK_EQUAL(addrman.size(), 1);
+        BOOST_CHECK_EQUAL(addrman.size(), (uint64_t)1);
         bool newOnly = true;
         CAddrInfo addr_ret3 = addrman.Select(newOnly);
         BOOST_CHECK_EQUAL(addr_ret3.ToString(), "250.1.1.1:8767");
@@ -175,7 +175,7 @@ BOOST_FIXTURE_TEST_SUITE(addrman_tests, BasicTestingSetup)
         // Test: Select from new with 1 addr in new.
         CService addr1 = ResolveService("250.1.1.1", 8767);
         addrman.Add(CAddress(addr1, NODE_NONE), source);
-        BOOST_CHECK_EQUAL(addrman.size(), 1);
+        BOOST_CHECK_EQUAL(addrman.size(), (uint64_t)1);
 
         bool newOnly = true;
         CAddrInfo addr_ret1 = addrman.Select(newOnly);
@@ -183,14 +183,14 @@ BOOST_FIXTURE_TEST_SUITE(addrman_tests, BasicTestingSetup)
 
         // Test: move addr to tried, select from new expected nothing returned.
         addrman.Good(CAddress(addr1, NODE_NONE));
-        BOOST_CHECK_EQUAL(addrman.size(), 1);
+        BOOST_CHECK_EQUAL(addrman.size(), (uint64_t)1);
         CAddrInfo addr_ret2 = addrman.Select(newOnly);
         BOOST_CHECK_EQUAL(addr_ret2.ToString(), "[::]:0");
 
         CAddrInfo addr_ret3 = addrman.Select();
         BOOST_CHECK_EQUAL(addr_ret3.ToString(), "250.1.1.1:8767");
 
-        BOOST_CHECK_EQUAL(addrman.size(), 1);
+        BOOST_CHECK_EQUAL(addrman.size(), (uint64_t)1);
 
 
         // Add three addresses to new table.
@@ -215,7 +215,7 @@ BOOST_FIXTURE_TEST_SUITE(addrman_tests, BasicTestingSetup)
         addrman.Good(CAddress(addr7, NODE_NONE));
 
         // Test: 6 addrs + 1 addr from last test = 7.
-        BOOST_CHECK_EQUAL(addrman.size(), 7);
+        BOOST_CHECK_EQUAL(addrman.size(), (uint64_t)7);
 
         // Test: Select pulls from new and tried regardless of port number.
         std::set<uint16_t> ports;
@@ -223,7 +223,7 @@ BOOST_FIXTURE_TEST_SUITE(addrman_tests, BasicTestingSetup)
         {
             ports.insert(addrman.Select().GetPort());
         }
-        BOOST_CHECK_EQUAL(ports.size(), 3);
+        BOOST_CHECK_EQUAL(ports.size(), (uint64_t)3);
     }
 
     BOOST_AUTO_TEST_CASE(addrman_new_collisions_test)
@@ -234,7 +234,7 @@ BOOST_FIXTURE_TEST_SUITE(addrman_tests, BasicTestingSetup)
 
         CNetAddr source = ResolveIP("252.2.2.2");
 
-        BOOST_CHECK_EQUAL(addrman.size(), 0);
+        BOOST_CHECK_EQUAL(addrman.size(), (uint64_t)0);
 
         for (unsigned int i = 1; i < 18; i++)
         {
@@ -248,11 +248,11 @@ BOOST_FIXTURE_TEST_SUITE(addrman_tests, BasicTestingSetup)
         //Test: new table collision!
         CService addr1 = ResolveService("250.1.1.18");
         addrman.Add(CAddress(addr1, NODE_NONE), source);
-        BOOST_CHECK_EQUAL(addrman.size(), 17);
+        BOOST_CHECK_EQUAL(addrman.size(), (uint64_t)17);
 
         CService addr2 = ResolveService("250.1.1.19");
         addrman.Add(CAddress(addr2, NODE_NONE), source);
-        BOOST_CHECK_EQUAL(addrman.size(), 18);
+        BOOST_CHECK_EQUAL(addrman.size(), (uint64_t)18);
     }
 
     BOOST_AUTO_TEST_CASE(addrman_tried_collisions_test)
@@ -263,7 +263,7 @@ BOOST_FIXTURE_TEST_SUITE(addrman_tests, BasicTestingSetup)
 
         CNetAddr source = ResolveIP("252.2.2.2");
 
-        BOOST_CHECK_EQUAL(addrman.size(), 0);
+        BOOST_CHECK_EQUAL(addrman.size(), (uint64_t)0);
 
         for (unsigned int i = 1; i < 80; i++)
         {
@@ -278,11 +278,11 @@ BOOST_FIXTURE_TEST_SUITE(addrman_tests, BasicTestingSetup)
         //Test: tried table collision!
         CService addr1 = ResolveService("250.1.1.80");
         addrman.Add(CAddress(addr1, NODE_NONE), source);
-        BOOST_CHECK_EQUAL(addrman.size(), 79);
+        BOOST_CHECK_EQUAL(addrman.size(), (uint64_t)79);
 
         CService addr2 = ResolveService("250.1.1.81");
         addrman.Add(CAddress(addr2, NODE_NONE), source);
-        BOOST_CHECK_EQUAL(addrman.size(), 80);
+        BOOST_CHECK_EQUAL(addrman.size(), (uint64_t)80);
     }
 
     BOOST_AUTO_TEST_CASE(addrman_find_test)
@@ -291,7 +291,7 @@ BOOST_FIXTURE_TEST_SUITE(addrman_tests, BasicTestingSetup)
 
         CAddrManTest addrman;
 
-        BOOST_CHECK_EQUAL(addrman.size(), 0);
+        BOOST_CHECK_EQUAL(addrman.size(), (uint64_t)0);
 
         CAddress addr1 = CAddress(ResolveService("250.1.2.1", 8767), NODE_NONE);
         CAddress addr2 = CAddress(ResolveService("250.1.2.1", 9999), NODE_NONE);
@@ -326,7 +326,7 @@ BOOST_FIXTURE_TEST_SUITE(addrman_tests, BasicTestingSetup)
 
         CAddrManTest addrman;
 
-        BOOST_CHECK_EQUAL(addrman.size(), 0);
+        BOOST_CHECK_EQUAL(addrman.size(), (uint64_t)0);
 
         CAddress addr1 = CAddress(ResolveService("250.1.2.1", 8767), NODE_NONE);
         CNetAddr source1 = ResolveIP("250.1.2.1");
@@ -348,7 +348,7 @@ BOOST_FIXTURE_TEST_SUITE(addrman_tests, BasicTestingSetup)
 
         CAddrManTest addrman;
 
-        BOOST_CHECK_EQUAL(addrman.size(), 0);
+        BOOST_CHECK_EQUAL(addrman.size(), (uint64_t)0);
 
         CAddress addr1 = CAddress(ResolveService("250.1.2.1", 8767), NODE_NONE);
         CNetAddr source1 = ResolveIP("250.1.2.1");
@@ -357,9 +357,9 @@ BOOST_FIXTURE_TEST_SUITE(addrman_tests, BasicTestingSetup)
         addrman.Create(addr1, source1, &nId);
 
         // Test: Delete should actually delete the addr.
-        BOOST_CHECK_EQUAL(addrman.size(), 1);
+        BOOST_CHECK_EQUAL(addrman.size(), (uint64_t)1);
         addrman.Delete(nId);
-        BOOST_CHECK_EQUAL(addrman.size(), 0);
+        BOOST_CHECK_EQUAL(addrman.size(), (uint64_t)0);
         CAddrInfo *info2 = addrman.Find(addr1);
         BOOST_CHECK(info2 == nullptr);
     }
@@ -372,9 +372,9 @@ BOOST_FIXTURE_TEST_SUITE(addrman_tests, BasicTestingSetup)
 
         // Test: Sanity check, GetAddr should never return anything if addrman
         //  is empty.
-        BOOST_CHECK_EQUAL(addrman.size(), 0);
+        BOOST_CHECK_EQUAL(addrman.size(), (uint64_t)0);
         std::vector<CAddress> vAddr1 = addrman.GetAddr();
-        BOOST_CHECK_EQUAL(vAddr1.size(), 0);
+        BOOST_CHECK_EQUAL(vAddr1.size(), (uint64_t)0);
 
         CAddress addr1 = CAddress(ResolveService("250.250.2.1", 8767), NODE_NONE);
         addr1.nTime = GetAdjustedTime(); // Set time so isTerrible = false
@@ -397,12 +397,12 @@ BOOST_FIXTURE_TEST_SUITE(addrman_tests, BasicTestingSetup)
         addrman.Add(addr5, source1);
 
         // GetAddr returns 23% of addresses, 23% of 5 is 1 rounded down.
-        BOOST_CHECK_EQUAL(addrman.GetAddr().size(), 1);
+        BOOST_CHECK_EQUAL(addrman.GetAddr().size(), (uint64_t)1);
 
         // Test: Ensure GetAddr works with new and tried addresses.
         addrman.Good(CAddress(addr1, NODE_NONE));
         addrman.Good(CAddress(addr2, NODE_NONE));
-        BOOST_CHECK_EQUAL(addrman.GetAddr().size(), 1);
+        BOOST_CHECK_EQUAL(addrman.GetAddr().size(), (uint64_t)1);
 
         // Test: Ensure GetAddr still returns 23% when addrman has many addrs.
         for (unsigned int i = 1; i < (8 * 256); i++)
@@ -422,9 +422,9 @@ BOOST_FIXTURE_TEST_SUITE(addrman_tests, BasicTestingSetup)
 
         size_t percent23 = (addrman.size() * 23) / 100;
         BOOST_CHECK_EQUAL(vAddr.size(), percent23);
-        BOOST_CHECK_EQUAL(vAddr.size(), 461);
+        BOOST_CHECK_EQUAL(vAddr.size(), (uint64_t)461);
         // (Addrman.size() < number of addresses added) due to address collisions.
-        BOOST_CHECK_EQUAL(addrman.size(), 2006);
+        BOOST_CHECK_EQUAL(addrman.size(), (uint64_t)2006);
     }
 
 
@@ -470,7 +470,7 @@ BOOST_FIXTURE_TEST_SUITE(addrman_tests, BasicTestingSetup)
         }
         // Test: IP addresses in the same group (\16 prefix for IPv4) should
         //  never get more than 8 buckets
-        BOOST_CHECK_EQUAL(buckets.size(), 8);
+        BOOST_CHECK_EQUAL(buckets.size(), (uint64_t)8);
 
         buckets.clear();
         for (int j = 0; j < 255; j++)
@@ -483,7 +483,7 @@ BOOST_FIXTURE_TEST_SUITE(addrman_tests, BasicTestingSetup)
         }
         // Test: IP addresses in the different groups should map to more than
         //  8 buckets.
-        BOOST_CHECK_EQUAL(buckets.size(), 160);
+        BOOST_CHECK_EQUAL(buckets.size(), (uint64_t)160);
     }
 
     BOOST_AUTO_TEST_CASE(caddrinfo_get_new_bucket_test)
@@ -526,7 +526,7 @@ BOOST_FIXTURE_TEST_SUITE(addrman_tests, BasicTestingSetup)
         }
         // Test: IP addresses in the same group (\16 prefix for IPv4) should
         //  always map to the same bucket.
-        BOOST_CHECK_EQUAL(buckets.size(), 1);
+        BOOST_CHECK_EQUAL(buckets.size(), (uint64_t)1);
 
         buckets.clear();
         for (int j = 0; j < 4 * 255; j++)
